Eerste commit!

+++ module-01/feestlunch.py
@@ -14,10 +14,10 @@ prijsst = stokbrood * stokbroodaantal
 prijszonderkorting = prijscr + prijsst
 
 kortingsbonnen = input ('hoeveel cent korting heb je in totaal? ')
-kortingsbonnen = float (kortingsbonnen)
+kortingsbonnen = int (kortingsbonnen)
 
 totaalprijs = (prijszonderkorting - kortingsbonnen) 
-totaalprijs = float (totaalprijs)
+totaalprijs = int (totaalprijs)
 
 print('De feestlunch kost je bij de bakker ', totaalprijs  ,' cent voor de ', croissantaantal ,' croissantjes en de ', stokbroodaantal ,' stokbroden als je ', kortingsbonnen  ,' cent korting hebt!')
